requirements.txt
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
google==3.0.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-vision==2.6.3
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
gspread==5.0.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
oauth2client==4.1.3
oauthlib==3.1.1
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pyparsing==3.0.6
python-dotenv==0.21.0
python-stdnum==1.17
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
soupsieve==2.3.1
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
